requirements.txt
Django==1.8.4
argparse==1.2.1
django-extensions==1.5.5
factory-boy==2.5.2
psycopg2==2.6.1
six==1.9.0
wsgiref==0.1.2
